test: Move tests folder out to be sibling of src/ directory
